feat(map): improve device map for multiple locations
This changes the behaviour of the device map so that
it no longer zooms to a level that includes all device
locations, but only zooms on the clicked on.

This was changed because some devices can have old,
far distant locations (e.g. single-cell) which no longer
is relevant.

Now the device map centers and zooms on the most recent
location and remembers which location is supposed to be
displayed (by source).

This also adds a button to center the device map on each
individual location source.

Finally, the tolerance for the GeoJSON object that renders
the different location sources is lowered so that the
hexagon is displayed at any zoom level.

Fixes #381
See https://github.com/hello-nrfcloud/map/discussions/378